+++ src/main/java/module/playerOverview/SpielerTrainingsVergleichsPanel.java
@@ -11,6 +11,7 @@ import core.model.HOVerwaltung;
 import core.model.UserParameter;
 import core.model.player.Player;
 import java.awt.BorderLayout;
+import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.Serial;
 import java.util.ArrayList;
@@ -18,12 +19,13 @@ import java.util.List;
 import javax.swing.*;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
+import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 
 
-
 /**
- * Hier kann eingestellt werden, mit welchem HRF die aktuelle Mannschaft verglichen werden soll
+ * Panel that displays the list of stored HRFs, which can be selected to perform
+ * comparisons with current team.
  */
 public class SpielerTrainingsVergleichsPanel extends ImagePanel
     implements Refreshable, ListSelectionListener, ActionListener {
@@ -40,7 +42,7 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
     //~ Instance fields ----------------------------------------------------------------------------
 
     private final JButton m_jbLoeschen = new JButton(HOVerwaltung.instance().getLanguageString("ls.button.delete"));
-    private final JList m_jlHRFs = new JList();
+    private final JList<CBItem> m_jlHRFs = new JList<>();
     private final List<ChangeListener> changeListeners = new ArrayList<>();
 
     //~ Constructors -------------------------------------------------------------------------------
@@ -66,7 +68,7 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
         return vergleichsPlayer;
     }
 
-    public final void actionPerformed(java.awt.event.ActionEvent actionEvent) {
+    public final void actionPerformed(ActionEvent actionEvent) {
         final var hrfs = m_jlHRFs.getSelectedValuesList();
         StringBuilder deleteInfoText = new StringBuilder(HOVerwaltung.instance().getLanguageString("ls.button.delete"));
 
@@ -78,7 +80,7 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
 
         var MAX_NUMBER_OF_FILENAMES_IN_DELETEINFOTEXT = 11;
         int i=0;
-        for (var hrf : hrfs){
+        for (CBItem hrf : hrfs){
             deleteInfoText.append("\n");
             if ( i++ < MAX_NUMBER_OF_FILENAMES_IN_DELETEINFOTEXT){
                 deleteInfoText.append(hrf.toString());
@@ -93,15 +95,15 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
 				HOVerwaltung.instance().getLanguageString("confirmation.title"), JOptionPane.YES_NO_OPTION);
 
         if (value == JOptionPane.OK_OPTION) {
-            for (Object hrf : hrfs) {
-                DBManager.instance().deleteHRF(((CBItem) hrf).getId());
+            for (CBItem hrf : hrfs) {
+                DBManager.instance().deleteHRF(hrf.getId());
             }
 
             loadHRFListe(false);
-            vergleichsPlayer.clear(); // .removeAllElements();
+            vergleichsPlayer.clear();
 
-            //Nur manuelles Update der Tabelle, kein reInit, damit die Sortierung bleibt.
-            HOMainFrame.instance().getSpielerUebersichtPanel().refreshHRFVergleich();
+            // Manual update of the table, so no reInit to keep the current sorting.
+            HOMainFrame.instance().getSpielerUebersichtPanel().refreshHRFComparison();
         }
     }
 
@@ -120,10 +122,10 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
     /**
      * Handle valueChanged() events.
      */
-	public final void valueChanged(javax.swing.event.ListSelectionEvent listSelectionEvent) {
+	public final void valueChanged(ListSelectionEvent listSelectionEvent) {
 		// Markierung vorhanden
 		if (m_jlHRFs.getSelectedValue() != null) {
-            hrfId = ((CBItem) m_jlHRFs.getSelectedValue()).getId();
+            hrfId = m_jlHRFs.getSelectedValue().getId();
 			vergleichsPlayer = DBManager.instance().getSpieler(hrfId);
 			vergleichsMarkierung = true;
 
@@ -141,7 +143,7 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
         fireChangeEvent(changeEvent);
 
 		// Manual update of the table, so no reInit to keep the current sorting.
-		HOMainFrame.instance().getSpielerUebersichtPanel().refreshHRFVergleich();
+		HOMainFrame.instance().getSpielerUebersichtPanel().refreshHRFComparison();
 	}
 
     /**
@@ -201,7 +203,11 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
 
         m_jlHRFs.removeListSelectionListener(this);
 
-        final Object letzteMarkierung = m_jlHRFs.getSelectedValue();
+        Object lastSelectedEntry = m_jlHRFs.getSelectedValue();
+        if (lastSelectedEntry == null && m_jlHRFs.getModel() instanceof DefaultListModel<CBItem> && m_jlHRFs.getModel().getSize() > 0) {
+            // Memorize first entry
+            lastSelectedEntry = m_jlHRFs.getModel().getElementAt(0);
+        }
 
         DefaultListModel listmodel;
 
@@ -212,25 +218,22 @@ public class SpielerTrainingsVergleichsPanel extends ImagePanel
             listmodel = new DefaultListModel();
         }
 
-        for (CBItem cbItem : hrfListe) {
-            listmodel.addElement(cbItem);
-        }
-
+        listmodel.addAll(hrfListe);
         m_jlHRFs.setModel(listmodel);
 
-        //Bei der Initialisierung noch keinen Vergleich anzeigen!
+        // Don't show comparison during init phase.
         if (!init) {
             m_jlHRFs.addListSelectionListener(this);
         }
 
-        if (letzteMarkierung != null) {
-            m_jlHRFs.setSelectedValue(letzteMarkierung, true);
+        if (lastSelectedEntry != null) {
+            m_jlHRFs.setSelectedValue(lastSelectedEntry, true);
         } else if ((listmodel.size() > 1) && !init) {
-            //Sonst das 2. HRF markieren, wenn es nicht direkt nach dem Start ist!
+            // Select the second HRF when not right after start
+            // (presumably to select and display comparison after a new HRF download)
             m_jlHRFs.setSelectedIndex(1);
         }
 
-        //Beim Initialisieren hier den Listener hinzufügen
         if (init) {
             m_jlHRFs.addListSelectionListener(this);
         }