+++ src/main/resources/release_notes.md
@@ -11,6 +11,7 @@
 ### Database
 * Fix `unexpected token` issue with Transfer table query (#1965)
 * Fix potential NPE when using a prepared statement (#1941)
+* Improve Database Cleanup tool by adding additional match types (#1587)
 * Repairing a failed V5 database upgrade (#1941)
 
 ### Squad
@@ -31,7 +32,8 @@
 ### Lineup
 * Fix missing player id column in lineup assistant's player table (#1930)
 
-### Statistics
+### Special events
+* Refactoring special events table layout (#816)
 
 ### Transfer
 * Refactoring of the transfer module. Calculation of transfer fee income. (#245)
@@ -49,6 +51,7 @@
 * Increase effect of youth friendly match training (#1950, #1994)
 * Modify youth module layout (#1449)
 * Set initial youth module view layout (#1558)
+* Display specialty icons in youth player tables (#1999)
 
 ### Misc
 * Add a Linux-friendly default theme, called “Gnome.”
@@ -56,6 +59,8 @@
 * Reset progress bar when action is finished (#1955)
 * Upgrade Gradle to version 8.5
 * Fix issue with startup post-installation (#2002)
+* Fix NPE when saving preferences on un-managed HO install (#1992)
+  HO now downloads the update in the browser when HO install is un-managed.
 
 ## Translations